move the functions that are create and drop tables
